[Fixes #4051] Split OutputFormatter into OutputFormatter and TextOutputFormatter
