+++ samples/FormatFilterSample.Web/CustomFormatter.cs
@@ -9,7 +9,7 @@ using Microsoft.Net.Http.Headers;
 
 namespace FormatFilterSample.Web
 {
-    public class CustomFormatter : OutputFormatter
+    public class CustomFormatter : TextOutputFormatter
     {
         public CustomFormatter(string contentType)
         {
@@ -38,7 +38,7 @@ namespace FormatFilterSample.Web
             return false;
         }
 
-        public override async Task WriteResponseBodyAsync(OutputFormatterWriteContext context)
+        public override async Task WriteResponseBodyAsync(OutputFormatterWriteContext context, Encoding selectedEncoding)
         {
             var response = context.HttpContext.Response;
             await response.WriteAsync(context.Object.ToString());
