+++ src/Microsoft.AspNetCore.Mvc.Core/Resources.resx
@@ -334,4 +334,10 @@
   <data name="TextInputFormatter_SupportedEncodingsMustNotBeEmpty" xml:space="preserve">
     <value>The list of '{0}' must not be empty. Add at least one supported encoding.</value>
   </data>
+  <data name="TextOutputFormatter_SupportedEncodingsMustNotBeEmpty" xml:space="preserve">
+    <value>The list of '{0}' must not be empty. Add at least one supported encoding.</value>
+  </data>
+  <data name="TextOutpurFormatter_WriteResponseBodyAsynNotSupported" xml:space="preserve">
+    <value>'{0}' is not supported by '{1}'. Use '{2}' instead.</value>
+  </data>
 </root>
