+++ test/Microsoft.AspNetCore.Mvc.Core.Test/Formatters/OutputFormatterTests.cs
@@ -18,228 +18,6 @@ namespace Microsoft.AspNetCore.Mvc.Formatters
 {
     public class OutputFormatterTests
     {
-        public static IEnumerable<object[]> SelectResponseCharacterEncodingData
-        {
-            get
-            {
-                // string acceptEncodings, string requestEncoding, string[] supportedEncodings, string expectedEncoding
-                yield return new object[] { "", new string[] { "utf-8", "utf-16" }, "utf-8" };
-
-                yield return new object[] { "utf-8", new string[] { "utf-8", "utf-16" }, "utf-8" };
-                yield return new object[] { "utf-16", new string[] { "utf-8", "utf-16" }, "utf-16" };
-                yield return new object[] { "utf-16; q=0.5", new string[] { "utf-8", "utf-16" }, "utf-16" };
-
-                yield return new object[] { "utf-8; q=0.0", new string[] { "utf-8", "utf-16" }, "utf-8" };
-                yield return new object[] { "utf-8; q=0.0, utf-16; q=0.0", new string[] { "utf-8", "utf-16" }, "utf-8" };
-
-                yield return new object[] { "*; q=0.0", new string[] { "utf-8", "utf-16" }, "utf-8" };
-            }
-        }
-
-        [Theory]
-        [MemberData(nameof(SelectResponseCharacterEncodingData))]
-        public void SelectResponseCharacterEncoding_SelectsEncoding(
-            string acceptCharsetHeaders,
-            string[] supportedEncodings,
-            string expectedEncoding)
-        {
-            // Arrange
-            var httpContext = new Mock<HttpContext>();
-            var httpRequest = new DefaultHttpContext().Request;
-            httpRequest.Headers[HeaderNames.AcceptCharset] = acceptCharsetHeaders;
-            httpRequest.Headers[HeaderNames.Accept] = "application/acceptCharset";
-            httpContext.SetupGet(o => o.Request).Returns(httpRequest);
-
-            var formatter = new TestOutputFormatter();
-            foreach (string supportedEncoding in supportedEncodings)
-            {
-                formatter.SupportedEncodings.Add(Encoding.GetEncoding(supportedEncoding));
-            }
-
-            var context = new OutputFormatterWriteContext(
-                httpContext.Object,
-                new TestHttpResponseStreamWriterFactory().CreateWriter,
-                typeof(string),
-                "someValue")
-            {
-                ContentType = new StringSegment(httpRequest.Headers[HeaderNames.Accept]),
-            };
-
-            // Act
-            var actualEncoding = formatter.SelectCharacterEncoding(context);
-
-            // Assert
-            Assert.Equal(Encoding.GetEncoding(expectedEncoding), actualEncoding);
-        }
-
-        [Theory]
-        [InlineData("application/json; charset=utf-16", "application/json; charset=utf-32")]
-        [InlineData("application/json; charset=utf-16; format=indent", "application/json; charset=utf-32; format=indent")]
-        public void WriteResponse_OverridesCharset_IfDifferentFromContentTypeCharset(
-            string contentType,
-            string expectedContentType)
-        {
-            // Arrange
-            var formatter = new Mock<OutputFormatter>();
-
-            formatter
-                .Setup(f => f.SelectCharacterEncoding(It.IsAny<OutputFormatterWriteContext>()))
-                .Returns(Encoding.UTF32);
-
-            var formatterContext = new OutputFormatterWriteContext(
-                new DefaultHttpContext(),
-                new TestHttpResponseStreamWriterFactory().CreateWriter,
-                objectType: null,
-                @object: null)
-            {
-                ContentType = new StringSegment(contentType),
-            };
-
-            // Act
-            formatter.Object.WriteAsync(formatterContext);
-
-            // Assert
-            Assert.Equal(new StringSegment(expectedContentType), formatterContext.ContentType);
-        }
-
-        [Fact]
-        public void WriteResponse_GetMediaTypeWithCharsetReturnsMediaTypeFromCache_IfEncodingIsUtf8()
-        {
-            // Arrange
-            var formatter = new TestOutputFormatter();
-
-            var formatterContext = new OutputFormatterWriteContext(
-                new DefaultHttpContext(),
-                new TestHttpResponseStreamWriterFactory().CreateWriter,
-                objectType: null,
-                @object: null)
-            {
-                ContentType = new StringSegment("application/json"),
-            };
-
-            formatter.SupportedMediaTypes.Add("application/json");
-            formatter.SupportedEncodings.Add(Encoding.UTF8);
-
-            // Act
-            formatter.WriteAsync(formatterContext);
-            var firstContentType = formatterContext.ContentType;
-
-            formatterContext.ContentType = new StringSegment("application/json");
-
-            formatter.WriteAsync(formatterContext);
-            var secondContentType = formatterContext.ContentType;
-
-            // Assert
-            Assert.Same(firstContentType.Buffer, secondContentType.Buffer);
-        }
-
-        [Fact]
-        public void WriteResponse_GetMediaTypeWithCharsetReplacesCharset_IfDifferentThanEncoding()
-        {
-            // Arrange
-            var formatter = new TestOutputFormatter();
-
-            var formatterContext = new OutputFormatterWriteContext(
-                new DefaultHttpContext(),
-                new TestHttpResponseStreamWriterFactory().CreateWriter,
-                objectType: null,
-                @object: null)
-            {
-                ContentType = new StringSegment("application/json; charset=utf-7"),
-            };
-
-            formatter.SupportedMediaTypes.Add("application/json");
-            formatter.SupportedEncodings.Add(Encoding.UTF8);
-
-            // Act
-            formatter.WriteAsync(formatterContext);
-
-            // Assert
-            Assert.Equal(new StringSegment("application/json; charset=utf-8"), formatterContext.ContentType);
-        }
-
-        [Fact]
-        public void WriteResponse_GetMediaTypeWithCharsetReturnsSameString_IfCharsetEqualToEncoding()
-        {
-            // Arrange
-            var formatter = new TestOutputFormatter();
-
-            var contentType = "application/json; charset=utf-16";
-            var formatterContext = new OutputFormatterWriteContext(
-                new DefaultHttpContext(),
-                new TestHttpResponseStreamWriterFactory().CreateWriter,
-                objectType: null,
-                @object: null)
-            {
-                ContentType = new StringSegment(contentType),
-            };
-
-            formatter.SupportedMediaTypes.Add("application/json");
-            formatter.SupportedEncodings.Add(Encoding.Unicode);
-
-            // Act
-            formatter.WriteAsync(formatterContext);
-
-            // Assert
-            Assert.Same(contentType, formatterContext.ContentType.Buffer);
-        }
-
-        [Fact]
-        public void WriteResponseContentHeaders_NoSupportedEncodings_NoEncodingIsSet()
-        {
-            // Arrange
-            var formatter = new TestOutputFormatter();
-
-            var testContentType = new StringSegment("text/json");
-
-            formatter.SupportedEncodings.Clear();
-            formatter.SupportedMediaTypes.Clear();
-            formatter.SupportedMediaTypes.Add(new MediaTypeHeaderValue("text/json"));
-
-            var context = new OutputFormatterWriteContext(
-                new DefaultHttpContext(),
-                new TestHttpResponseStreamWriterFactory().CreateWriter,
-                objectType: null,
-                @object: null)
-            {
-                ContentType = testContentType,
-            };
-
-            // Act
-            formatter.WriteResponseHeaders(context);
-
-            // Assert
-            Assert.Null(MediaTypeHeaderValue.Parse(context.ContentType.Value).Encoding);
-            Assert.Equal(testContentType, context.ContentType);
-
-            // If we had set an encoding, it would be part of the content type header
-            Assert.Equal(MediaTypeHeaderValue.Parse(testContentType.Value), context.HttpContext.Response.GetTypedHeaders().ContentType);
-        }
-
-        [Fact]
-        public async Task WriteResponseHeaders_ClonesMediaType()
-        {
-            // Arrange
-            var formatter = new PngImageFormatter();
-            formatter.SupportedMediaTypes.Clear();
-            var mediaType = new MediaTypeHeaderValue("image/png");
-            formatter.SupportedMediaTypes.Add(mediaType);
-
-            var context = new OutputFormatterWriteContext(
-                new DefaultHttpContext(),
-                new TestHttpResponseStreamWriterFactory().CreateWriter,
-                objectType: null,
-                @object: null);
-
-            // Act
-            await formatter.WriteAsync(context);
-
-            // Assert
-            Assert.NotSame(mediaType, context.ContentType);
-            Assert.Null(mediaType.Charset);
-            Assert.Equal("image/png; charset=utf-8", context.ContentType.ToString());
-        }
-
         [Fact]
         public void CanWriteResult_ForNullContentType_UsesFirstEntryInSupportedContentTypes()
         {
@@ -397,42 +175,7 @@ namespace Microsoft.AspNetCore.Mvc.Formatters
         {
             public TestOutputFormatter()
             {
-                SupportedMediaTypes.Add(MediaTypeHeaderValue.Parse("application/acceptCharset"));
-            }
-
-            public override Task WriteResponseBodyAsync(OutputFormatterWriteContext context)
-            {
-                return Task.FromResult(true);
-            }
-        }
-
-        private class DoesNotSetContext : OutputFormatter
-        {
-            public DoesNotSetContext()
-            {
-                SupportedMediaTypes.Add(MediaTypeHeaderValue.Parse("application/doesNotSetContext"));
-                SupportedEncodings.Add(Encoding.Unicode);
-            }
-
-            public override bool CanWriteResult(OutputFormatterCanWriteContext context)
-            {
-                // Do not set the selected media Type.
-                // The WriteResponseHeaders should do it for you.
-                return true;
-            }
-
-            public override Task WriteResponseBodyAsync(OutputFormatterWriteContext context)
-            {
-                return Task.FromResult(true);
-            }
-        }
-
-        private class PngImageFormatter : OutputFormatter
-        {
-            public PngImageFormatter()
-            {
-                SupportedMediaTypes.Add(MediaTypeHeaderValue.Parse("image/png"));
-                SupportedEncodings.Add(Encoding.UTF8);
+                SupportedMediaTypes.Add("application/acceptCharset");
             }
 
             public override Task WriteResponseBodyAsync(OutputFormatterWriteContext context)
