+++ test/WebSites/BasicWebSite/Formatters/CustomFormatter.cs
@@ -9,7 +9,7 @@ using Microsoft.Net.Http.Headers;
 
 namespace BasicWebSite.Formatters
 {
-    public class CustomFormatter : OutputFormatter
+    public class CustomFormatter : TextOutputFormatter
     {
         public string ContentType { get; private set; }
 
@@ -33,7 +33,7 @@ namespace BasicWebSite.Formatters
             return false;
         }
 
-        public override async Task WriteResponseBodyAsync(OutputFormatterWriteContext context)
+        public override async Task WriteResponseBodyAsync(OutputFormatterWriteContext context, Encoding selectedEncoding)
         {
             var response = context.HttpContext.Response;
             response.ContentType = ContentType + ";charset=utf-8";
